Fixed typedefs

+++ src/globals.h
@@ -22,11 +22,11 @@
 
 #define NIL 0
 
-typedef short unsigned char u8;
+typedef unsigned char u8;
 typedef short unsigned int u16;
 typedef unsigned int u32;
 typedef long unsigned int u64;
-typedef char u8;
+typedef char i8;
 typedef short int i16;
 typedef int i32;
 typedef long int i64;
